requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
Flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-limiter==3.7.0

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30

# ── Caching ───────────────────────────────────────────────────────────────────
redis==5.0.4

# ── Security ──────────────────────────────────────────────────────────────────
werkzeug==3.0.3
bcrypt==4.1.3

# ── ML / Forecasting ─────────────────────────────────────────────────────────
prophet==1.1.5
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4

# ── HTTP / Utilities ─────────────────────────────────────────────────────────
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4

# ── PDF Report Generation ────────────────────────────────────────────────────
reportlab==4.2.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-flask==1.3.0
pytest-cov==5.0.0
